refactor to combineReducer implementation
